+++ ThirdPartyAdapters/unity/unity/src/main/java/com/google/ads/mediation/unity/UnityRewardedAd.java
@@ -44,6 +44,7 @@ import com.unity3d.ads.UnityAds.UnityAdsLoadError;
 import com.unity3d.ads.UnityAds.UnityAdsShowError;
 import com.unity3d.ads.UnityAdsLoadOptions;
 import com.unity3d.ads.UnityAdsShowOptions;
+import com.unity3d.ads.metadata.MetaData;
 import java.util.UUID;
 
 /**
@@ -235,9 +236,8 @@ public class UnityRewardedAd implements MediationRewardedAd {
                   + "and can now load rewarded ad with placement ID: %s",
               gameId, placementId);
       Log.d(TAG, logMessage);
-      // TODO(b/280861464): Add setCoppa test when loading ad
-      UnityAdsAdapterUtils.setCoppa(
-          MobileAds.getRequestConfiguration().getTagForChildDirectedTreatment(), context);
+      UnityAdsAdapterUtils.setUnityAdsPrivacy(
+          MobileAds.getRequestConfiguration(), new MetaData(context));
 
       objectId = UUID.randomUUID().toString();
       UnityAdsLoadOptions unityAdsLoadOptions =