+++ ThirdPartyAdapters/unity/unity/src/test/kotlin/com/google/ads/mediation/unity/UnityAdsAdapterUtilsTest.kt
@@ -4,15 +4,20 @@ import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.android.gms.ads.AdSize
+import com.google.android.gms.ads.MobileAds
+import com.google.android.gms.ads.RequestConfiguration
 import com.google.common.truth.Truth.assertThat
 import com.unity3d.ads.UnityAds.UnityAdsInitializationError
 import com.unity3d.ads.UnityAds.UnityAdsLoadError
 import com.unity3d.ads.UnityAds.UnityAdsShowError
+import com.unity3d.ads.metadata.MetaData
 import com.unity3d.services.banners.BannerErrorCode
 import com.unity3d.services.banners.BannerErrorInfo
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.kotlin.mock
+import org.mockito.kotlin.verify
 
 /** Unit tests for [UnityAdsAdapterUtils]. */
 @RunWith(AndroidJUnit4::class)
@@ -20,6 +25,18 @@ class UnityAdsAdapterUtilsTest {
 
   private val context: Context = ApplicationProvider.getApplicationContext()
   private var bannerErrorInfo: BannerErrorInfo = mock()
+
+  @Before
+  fun setUp() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(
+          RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
+        )
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
+  }
 
   @Test
   fun getMediationErrorCode_withBannerErrorInfo_returnsCorrectValueForUnknownEnum() {
@@ -236,5 +253,93 @@ class UnityAdsAdapterUtilsTest {
     assertThat(adError.getCode()).isEqualTo(200)
     assertThat(adError.getMessage()).isEqualTo("Description")
     assertThat(adError.getDomain()).isEqualTo(UnityMediationAdapter.SDK_ERROR_DOMAIN)
+  }
+
+  @Test
+  fun setUnityAdsPrivacy_withTFCDTrue_commitsNonBehavioralMetaDataTrue() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE)
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
+    val mockMetaData = mock<MetaData>()
+
+    UnityAdsAdapterUtils.setUnityAdsPrivacy(requestConfiguration, mockMetaData)
+
+    verify(mockMetaData).set("user.nonbehavioral", true)
+    verify(mockMetaData).commit()
+  }
+
+  @Test
+  fun setUnityAdsPrivacy_withTFUATrue_commitsNonBehavioralMetaDataTrue() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(
+          RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
+        )
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_TRUE)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
+    val mockMetaData = mock<MetaData>()
+
+    UnityAdsAdapterUtils.setUnityAdsPrivacy(requestConfiguration, mockMetaData)
+
+    verify(mockMetaData).set("user.nonbehavioral", true)
+    verify(mockMetaData).commit()
+  }
+
+  @Test
+  fun setUnityAdsPrivacy_withTFCDFalse_commitsNonBehavioralMetaDataFalse() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(
+          RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_FALSE
+        )
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
+    val mockMetaData = mock<MetaData>()
+
+    UnityAdsAdapterUtils.setUnityAdsPrivacy(requestConfiguration, mockMetaData)
+
+    verify(mockMetaData).set("user.nonbehavioral", false)
+    verify(mockMetaData).commit()
+  }
+
+  @Test
+  fun setUnityAdsPrivacy_withTFUAFalse_commitsNonBehavioralMetaDataFalse() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(
+          RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
+        )
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_FALSE)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
+    val mockMetaData = mock<MetaData>()
+
+    UnityAdsAdapterUtils.setUnityAdsPrivacy(requestConfiguration, mockMetaData)
+
+    verify(mockMetaData).set("user.nonbehavioral", false)
+    verify(mockMetaData).commit()
+  }
+
+  @Test
+  fun setUnityAdsPrivacy_withTFUAUnspecifiedAndTFCDUnspecified_commitsNonBehavioralMetaDataFalse() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(
+          RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
+        )
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_FALSE)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
+    val mockMetaData = mock<MetaData>()
+
+    UnityAdsAdapterUtils.setUnityAdsPrivacy(requestConfiguration, mockMetaData)
+
+    verify(mockMetaData).set("user.nonbehavioral", false)
+    verify(mockMetaData).commit()
   }
 }